initial pass to split Util and some of Constants into separate common files
