+++ src/Constants.h
@@ -15,31 +15,6 @@
 // file ids
 #define ALL_FILES -1
 
-// region ids
-#define CUBE_REGION_ID -2
-#define IMAGE_REGION_ID -1
-#define CURSOR_REGION_ID 0
-#define ALL_REGIONS -10
-
-// x axis
-#define ALL_X -2
-
-// y axis
-#define ALL_Y -2
-
-// z axis
-#define DEFAULT_Z 0
-#define CURRENT_Z -1
-#define ALL_Z -2
-
-// stokes
-#define DEFAULT_STOKES 0
-#define CURRENT_STOKES -1
-
-// raster image data
-#define TILE_SIZE 256
-#define CHUNK_SIZE 512
-
 // histograms
 #define AUTO_BIN_SIZE -1
 #define HISTOGRAM_CANCEL -1.0
@@ -50,10 +25,6 @@
 #define TARGET_DELTA_TIME 50 // milliseconds
 #define TARGET_PARTIAL_CURSOR_TIME 500
 #define TARGET_PARTIAL_REGION_TIME 1000
-
-// file list
-#define FILE_LIST_FIRST_PROGRESS_AFTER_SECS 5
-#define FILE_LIST_PROGRESS_INTERVAL_SECS 2
 
 // uWebSockets setting
 #define MAX_BACKPRESSURE 256 * 1024 * 1024
