+++ src/FileList/FileInfoLoader.cc
@@ -13,6 +13,9 @@
 #include <casacore/casa/OS/Directory.h>
 #include <casacore/casa/OS/File.h>
 
+#include "Util/Casacore.h"
+#include "Util/File.h"
+
 FileInfoLoader::FileInfoLoader(const std::string& filename) : _filename(filename) {
     _type = GetCartaFileType(filename);
 }
