+++ src/FileList/FileListHandler.h
@@ -47,7 +47,7 @@ private:
     // ICD: File/Region list response
     void GetFileList(CARTA::FileListResponse& file_list, std::string folder, ResultMsg& result_msg, bool region_list = false);
 
-    bool FillRegionFileInfo(CARTA::FileInfo& file_info, const string& filename, CARTA::FileType type = CARTA::FileType::UNKNOWN);
+    bool FillRegionFileInfo(CARTA::FileInfo& file_info, const std::string& filename, CARTA::FileType type = CARTA::FileType::UNKNOWN);
     void GetRegionFileContents(std::string& full_name, std::vector<std::string>& file_contents);
     void GetRelativePath(std::string& folder);
     CARTA::FileType GetRegionType(const std::string& filename); // parse first line for CRTF or DS9
