+++ src/ImageData/CompressedFits.cc
@@ -10,6 +10,7 @@
 #include <cmath>
 
 #include "../Logger/Logger.h"
+#include "Util/String.h"
 
 #define FITS_BLOCK_SIZE 2880
 #define FITS_CARD_SIZE 80
