+++ src/ImageData/FileLoader.cc
@@ -12,6 +12,8 @@
 #include <casacore/lattices/Lattices/MaskedLatticeIterator.h>
 
 #include "../Logger/Logger.h"
+#include "Util/File.h"
+#include "Util/Image.h"
 
 #include "CasaLoader.h"
 #include "CompListLoader.h"
