+++ src/Region/Region.h
@@ -26,6 +26,13 @@
 
 #define DEFAULT_VERTEX_COUNT 1000
 
+inline std::string RegionName(CARTA::RegionType type) {
+    std::unordered_map<CARTA::RegionType, std::string> region_names = {{CARTA::RegionType::POINT, "point"},
+        {CARTA::RegionType::LINE, "line"}, {CARTA::RegionType::POLYLINE, "polyline"}, {CARTA::RegionType::RECTANGLE, "rectangle"},
+        {CARTA::RegionType::ELLIPSE, "ellipse"}, {CARTA::RegionType::ANNULUS, "annulus"}, {CARTA::RegionType::POLYGON, "polygon"}};
+    return region_names[type];
+}
+
 struct RegionState {
     // struct used for region parameters
     int reference_file_id;
