+++ src/Region/RegionHandler.cc
@@ -18,6 +18,7 @@
 #include "../Constants.h"
 #include "../ImageStats/StatsCalculator.h"
 #include "../Logger/Logger.h"
+#include "Util/Image.h"
 
 #include "CrtfImportExport.h"
 #include "Ds9ImportExport.h"
