+++ src/Region/RegionImportExport.cc
@@ -12,6 +12,7 @@
 #include <imageanalysis/Annotations/AnnotationBase.h>
 
 #include "../Logger/Logger.h"
+#include "Util/String.h"
 
 using namespace carta;
