+++ src/Table/TableController.cc
@@ -10,6 +10,7 @@
 
 #include "Logger/Logger.h"
 #include "Timer/ListProgressReporter.h"
+#include "Util/File.h"
 
 #if defined(__APPLE__)
 #define st_mtim st_mtimespec
