+++ src/Util/File.h
@@ -7,17 +7,23 @@
 #ifndef CARTA_BACKEND__UTIL_FILE_H_
 #define CARTA_BACKEND__UTIL_FILE_H_
 
+#include "FileSystem.h"
+
 // Valid for little-endian only
 #define FITS_MAGIC_NUMBER 0x504D4953
 #define GZ_MAGIC_NUMBER 0x08088B1F
 #define HDF5_MAGIC_NUMBER 0x46444889
 #define XML_MAGIC_NUMBER 0x6D783F3C
 
+// file list
+#define FILE_LIST_FIRST_PROGRESS_AFTER_SECS 5
+#define FILE_LIST_PROGRESS_INTERVAL_SECS 2
+
 uint32_t GetMagicNumber(const std::string& filename);
 bool IsCompressedFits(const std::string& filename);
 
-// stokes types and value conversion
-int GetStokesValue(const CARTA::StokesType& stokes_type);
-CARTA::StokesType GetStokesType(int stokes_value);
+// directory functions
+int GetNumItems(const std::string& path);
+fs::path SearchPath(std::string filename);
 
 #endif // CARTA_BACKEND__UTIL_FILE_H_
