--- src/Util/Path.cc
@@ -1,57 +1,0 @@
-/* This file is part of the CARTA Image Viewer: https://github.com/CARTAvis/carta-backend
-   Copyright 2018, 2019, 2020, 2021 Academia Sinica Institute of Astronomy and Astrophysics (ASIAA),
-   Associated Universities, Inc. (AUI) and the Inter-University Institute for Data Intensive Astronomy (IDIA)
-   SPDX-License-Identifier: GPL-3.0-or-later
-*/
-
-#include "Path.h"
-
-bool FindExecutablePath(string& path) {
-    char path_buffer[PATH_MAX + 1];
-#ifdef __APPLE__
-    uint32_t len = sizeof(path_buffer);
-
-    if (_NSGetExecutablePath(path_buffer, &len) != 0) {
-        return false;
-    }
-#else
-    const int len = int(readlink("/proc/self/exe", path_buffer, PATH_MAX));
-
-    if (len == -1) {
-        return false;
-    }
-
-    path_buffer[len] = 0;
-#endif
-    path = path_buffer;
-    return true;
-}
-
-int GetNumItems(const string& path) {
-    try {
-        int counter = 0;
-        auto it = fs::directory_iterator(path);
-        for (const auto f : it) {
-            counter++;
-        }
-        return counter;
-    } catch (exception) {
-        return -1;
-    }
-}
-
-// quick alternative to bp::search_path that allows us to remove
-// boost:filesystem dependency
-fs::path SearchPath(std::string filename) {
-    std::string path(std::getenv("PATH"));
-    std::vector<std::string> path_strings;
-    SplitString(path, ':', path_strings);
-    for (auto& p : path_strings) {
-        fs::path base_path(p);
-        base_path /= filename;
-        if (fs::exists(base_path)) {
-            return base_path;
-        }
-    }
-    return fs::path();
-}
