+++ src/Util/String.cc
@@ -6,11 +6,13 @@
 
 #include "String.h"
 
-void SplitString(string& input, char delim, vector<string>& parts) {
+#include <sstream>
+
+void SplitString(std::string& input, char delim, std::vector<std::string>& parts) {
     // util to split input string into parts by delimiter
     parts.clear();
-    stringstream ss(input);
-    string item;
+    std::stringstream ss(input);
+    std::string item;
     while (getline(ss, item, delim)) {
         if (!item.empty()) {
             if (item.back() == '\r') {
