+++ src/Util/String.h
@@ -7,6 +7,9 @@
 #ifndef CARTA_BACKEND__UTIL_STRING_H_
 #define CARTA_BACKEND__UTIL_STRING_H_
 
+#include <string>
+#include <vector>
+
 // split input string into a vector of strings by delimiter
 void SplitString(std::string& input, char delim, std::vector<std::string>& parts);
