+++ src/Util/Token.h
@@ -7,6 +7,9 @@
 #ifndef CARTA_BACKEND__UTIL_TOKEN_H_
 #define CARTA_BACKEND__UTIL_TOKEN_H_
 
+#include <uWebSockets/HttpContext.h>
+#include <string>
+
 bool ValidateAuthToken(uWS::HttpRequest* http_request, const std::string& required_token);
 bool ConstantTimeStringCompare(const std::string& a, const std::string& b);
