+++ test/CommonTestUtilities.h
@@ -14,7 +14,7 @@
 #include <gtest/gtest.h>
 #include <spdlog/fmt/fmt.h>
 
-#include "Util.h"
+#include "App.h"
 
 #ifdef _BOOST_FILESYSTEM_
 #include <boost/filesystem.hpp>
@@ -23,8 +23,6 @@ namespace fs = boost::filesystem;
 #include <filesystem>
 namespace fs = std::filesystem;
 #endif
-
-using namespace carta;
 
 fs::path TestRoot();
 fs::path UserDirectory();
