+++ test/TestFitsTable.cc
@@ -6,7 +6,6 @@
 #include <gtest/gtest.h>
 
 #include "Table/Table.h"
-#include "Util.h"
 
 #include "CommonTestUtilities.h"
