+++ test/TestMoment.cc
@@ -8,7 +8,6 @@
 
 #include "Logger/Logger.h"
 #include "Moment/ImageMoments.h"
-#include "Util.h"
 
 #include <casacore/images/Images/PagedImage.h>
 #include <imageanalysis/ImageAnalysis/ImageMoments.h>
