+++ test/TestVoTable.cc
@@ -7,7 +7,6 @@
 #include <gtest/gtest.h>
 
 #include "Table/Table.h"
-#include "Util.h"
 
 #include "CommonTestUtilities.h"
